tanner/requirements.txt
django==1.11
django-extensions
python-ldap==3.0.0
